Remove needless (int) cast in bbp_get_forum_subforum_count
git-svn-id: https://bbpress.svn.wordpress.org/branches/plugin@2852 9866e705-20ec-0310-96e7-cbb4277adcfb

+++ bbp-includes/bbp-forum-template.php
@@ -892,7 +892,7 @@ function bbp_forum_subforum_count( $forum_id = 0 ) {
 	 */
 	function bbp_get_forum_subforum_count( $forum_id = 0 ) {
 		$forum_id    = bbp_get_forum_id( $forum_id );
-		$forum_count = (int) get_post_meta( $forum_id, '_bbp_forum_subforum_count', true );
+		$forum_count = get_post_meta( $forum_id, '_bbp_forum_subforum_count', true );
 
 		return apply_filters( 'bbp_get_forum_subforum_count', (int) $forum_count, $forum_id );
 	}
